# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Hotel Booking API",
    "version": "1.0.0",
    "title": "Hotel Booking API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },

  "tags": [
    {
      "name": "Products API",
      "description": "Test API"
    }
  ],
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },

    "schemas": {
      "HotelModel": {
        "type": "object",
        "properties": {
          "amenities": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "rating": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "register": {
        "type": "object",
        "properties": {
          "public_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          }
        }
      },
      "login": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/": {
      "post": {
        "tags": ["Add Data"],
        "summary": "Add Data",
        "requestBody": {
          "description": "Post new data to",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotelModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          }
        }
      },
      "get": {
        "tags": ["Get All Data"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get All Data",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": ["Add Data"],
        "summary": "Add Data",
        "requestBody": {
          "description": "Post new data to",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/register"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/register"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/register"
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["Add Data"],
        "summary": "Add Data",
        "requestBody": {
          "description": "Post new data to",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/register"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/register"
            }
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": ["Get All Search data from API"],
        "summary": "Get All data",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Search by title",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/HotelModel/properties/title"
            }
          },
          {
            "name": "location",
            "in": "query",
            "description": "search by location",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/HotelModel/properties/location"
            }
          },
          {
            "name": "amenities",
            "in": "query",
            "description": "Search by amenities",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/HotelModel/properties/amenities"
            }
          },

          {
            "name": "price1",
            "in": "query",
            "description": "Sort by price asc/desc",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/HotelModel/properties/price"
            }
          },
          {
            "name": "price",
            "in": "query",
            "description": "Search by price",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/HotelModel/properties/price"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          },
          "404": {
            "description": "NOT FOUND",
            "schema": {
              "$ref": "#/components/schemas/HotelModel"
            }
          }
        }
      }
    }
  }
}